Move preset methods from golem.client to golem.config
